+++ CHANGES.rst
@@ -15,13 +15,26 @@ Unreleased
   reduce false negatives by using approximate content matching,
   and make it possible to re-run the plugin for existing entries.
   (:issue:`202`)
-* Add :attr:`~Reader.after_feed_update_hooks`,
-  which allow running arbitrary actions for updated feeds.
+* Allow running arbitrary actions for updated feeds
+  via :attr:`~Reader.after_feed_update_hooks`.
   (:issue:`202`)
+
 * Allow :meth:`~Reader.mark_entry_as_read`
   and :meth:`~Reader.mark_entry_as_important`
   to mark an entry as unread/unimportant through a boolean flag.
   (:issue:`256`)
+
+* Record the time an entry is marked as read/important,
+  and make it available through :attr:`~Entry.read_modified` and
+  :attr:`~Entry.important_modified`.
+  Allow setting it to a user-provided value using the ``modified``
+  argument of :meth:`~Reader.mark_entry_as_read`
+  and :meth:`~Reader.mark_entry_as_important`.
+  (:issue:`254`)
+* Make :mod:`~reader.plugins.entry_dedupe` copy
+  :attr:`~Entry.read_modified` and :attr:`~Entry.important_modified`
+  from the duplicates to the new entry.
+  (:issue:`254`)
 
 * In the web application, allow marking an entry as "don't care"
   (read + unimportant explicitly set by the user) with a single button.